pyproject.toml
 [build-system]
 requires = ["setuptools>=68", "wheel"]
 build-backend = "setuptools.build_meta"

 [project]
 name = "rev-langgraph-example"
 version = "0.1.0"
 description = "LangGraph + Braintrust demo agent"
 requires-python = ">=3.11"
 dependencies = [
   "braintrust<0.6.0",
   "braintrust-langchain>=0.2.0",
   "langgraph>=0.2.0",
   "langchain>=0.2.0",
   "langchain-openai>=0.2.0",
   "langchain-community>=0.2.0",
   "langchain-text-splitters>=0.2.0",
   "faiss-cpu>=1.8.0",
   "tavily-python>=0.3.3",
   "fastapi>=0.110.0",
   "uvicorn>=0.29.0",
   "pydantic>=2.6.0",
   "python-dotenv>=1.0.1",
   "httpx>=0.27.0",
   "python-multipart",
   "pypdf>=4.0.0",
   "autoevals>=0.0.130",
]

 [project.optional-dependencies]
 dev = [
   "pytest>=7.4.0",
   "pytest-asyncio>=0.23.0",
 ]

 [tool.pytest.ini_options]
 testpaths = ["tests"]
 asyncio_mode = "auto"
